pyproject.toml
[project]
name = "teleop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.10.19"
dependencies = [
    "aiohttp>=3.13.2",
    "aiortc>=1.14.0",
    "av>=15.0.0,<=16.0.0",
    "cv2-enumerate-cameras>=1.3.1",
    "lerobot[all]<=0.4.0",
    "opencv-python>=4.12.0.88",
    "python-dotenv>=1.0.1",
]
